docs: near_events explained feat: `neardata` renamed to `near_events` for clarity
